fix: cleaning filter function
